+++ GCMFormTableView/GCMFormTableView/GCMItemSelectSection.m
@@ -17,7 +17,7 @@
 - (instancetype)initWithHeader:(NSAttributedString *)header
                         footer:(NSAttributedString *)footer
                     indexTitle:(NSString *)indexTitle
-            andSeparatorHeight:(float)height {
+            andSeparatorHeight:(CGFloat)height {
   self = [super init];
   if ( self ) {
     _items = [[NSMutableArray alloc] init];
